requirements.txt
click>=5.0
ruamel.yaml>=0.15